Move to mimalloc for rust.
