+++ .gitignore
@@ -11,3 +11,8 @@
 # csharp
 /bin
 /obj
+
+
+# Added by cargo
+
+/target
