+++ build-rs.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env sh
 
 mkdir -p build/rust
-rustc -O -o build/rust/almost_pseudo_random src/main/rust/almost_pseudo_random.rs
-rustc -O -o build/rust/rust_safer src/main/rust/rust_safer.rs
-rustc -O -o build/rust/rust_raw src/main/rust/rust_raw.rs
+cargo build --release --manifest-path src/main/rust/Cargo.toml
+find src/main/rust/target/release \
+  -maxdepth 1 \
+  -executable \
+  -type f \
+  -exec cp -f {} build/rust \;
