+++ src/main/rust/.gitignore
@@ -1,0 +1,2 @@
+target
+Cargo.lock
