+++ src/main/rust/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "rust"
+version = "0.1.0"
+authors = ["Sam <tritoke@protonmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+# compile mimalloc for speed not security
+mimalloc = { version = "~0.1.24", default-features = false }
