+++ src/main/rust/src/bin/almost_pseudo_random.rs
